make wait time for workflow startup configurable
At the moment we are waiting for a maximum of 6 seconds for the startup of the
triggered GitHub workflow. Sometimes the workflow takes longer and we get
failed runs even though the triggered action succeeds. Let us allow users
to configure the wait time to find a value that works for them.